Refactoring van_rossum_dist.
