+++ spykeutils/tests/test_signal_processing.py
@@ -13,6 +13,53 @@ import scipy as sp
 import spykeutils.signal_processing as sigproc
 
 
+class Test_searchsorted_pairwise(ut.TestCase):
+    def assert_array_tuple_equal(self, expected, actual):
+        self.assertEqual(len(expected), len(actual))
+        for e, a in zip(expected, actual):
+            assert_array_almost_equal(
+                e, a,
+                err_msg="Expected: {0}\nActual: {1}".format(expected, actual))
+
+    def test_works_with_lists(self):
+        a = [1, 3, 5, 8, 9, 11]
+        b = [2, 4, 6, 7, 10]
+        expected = ([-1, 0, 1, 3, 3, 4], [0, 1, 2, 2, 4])
+        actual = sigproc._searchsorted_pairwise(a, b)
+        self.assert_array_tuple_equal(expected, actual)
+
+    def test_works_with_array(self):
+        a = sp.array([1, 3, 5, 8, 9, 11])
+        b = sp.array([2, 4, 6, 7, 10])
+        expected = ([-1, 0, 1, 3, 3, 4], [0, 1, 2, 2, 4])
+        actual = sigproc._searchsorted_pairwise(a, b)
+        self.assert_array_tuple_equal(expected, actual)
+
+    def test_equal_items_in_second_list_are_considered_smaller(self):
+        a = [1, 3, 4]
+        b = [2, 3]
+        expected = ([-1, 1, 1], [0, 0])
+        actual = sigproc._searchsorted_pairwise(a, b)
+        self.assert_array_tuple_equal(expected, actual)
+
+    def test_works_with_one_list_empty(self):
+        a = []
+        b = [1, 2, 3]
+
+        expected = ([], [-1, -1, -1])
+        actual = sigproc._searchsorted_pairwise(a, b)
+        self.assert_array_tuple_equal(expected, actual)
+
+        expected = ([-1, -1, -1], [])
+        actual = sigproc._searchsorted_pairwise(b, a)
+        self.assert_array_tuple_equal(expected, actual)
+
+    def test_returns_empty_lists_if_both_lists_are_empty(self):
+        self.assert_array_tuple_equal(
+            ([], []),
+            sigproc._searchsorted_pairwise([], []))
+
+
 class TestCausalDecayingExpKernel(ut.TestCase):
     def setUp(self):
         self.kernel_size = 500 * pq.ms
